[refactor]Map画面をリファクタリング
Map画面が冗長だったため、コンポーネントを分割しました。
