+++ src/app/map/page.module.scss
@@ -20,20 +20,6 @@
     height: 100%;
     overflow: scroll;
   }
-
-  .infoWindowContent,
-  .selectedInfoWindowContent {
-    padding: 10px;
-  }
-
-  .hovered,
-  .selectedInfoWindowContent,
-  .travelInfoWindowContent,
-  .infoWindowContent:hover {
-    font-size: 20px;
-    font-weight: bold;
-    padding: 15px;
-  }
 }
 
 .container :global(.gm-style-iw) {
