+++ src/features/map/components/UserMarker.module.scss
@@ -1,0 +1,5 @@
+.infoWindow {
+  font-size: 20px;
+  font-weight: bold;
+  padding: 15px;
+}
